Fixed issue on model insert calls
Fixed cast issue (int -> long) that could potentially cause issues

+++ app/models/Demo.scala
@@ -145,7 +145,7 @@ object Demo {
           'codeurl -> demo.codeurl,
           'demourl -> demo.demourl,
           'description -> demo.description
-        ).as(int("id").single)
+        ).as(long("id").single)
 
         //add tags to the demo
         Tag.addToDemo(demo.tags, id)
